add to cart test convert to junit
